Make bashscaper opt-in with parameter

+++ Jenkinsfile
@@ -7,6 +7,7 @@ node {
     def skipNaisible = params.skipNaisible
     def naiscaperVersion = '31.0.0'
     def naiscaperDryRunVersion = '33.0.0'
+    def useBashscaper = params.useBashscaper
     def bashscaperVersion = '4.0.0'
     def naisplaterVersion = '6.0.0'
     def kubectlImageTag = 'v1.12.3'
@@ -129,46 +130,50 @@ node {
         }
 
         stage("update nais platform apps") {
-            sh("docker run --rm -v `pwd`/nais-platform-apps:/root/nais-platform-apps -v `pwd`/${clusterName}:/root/.kube navikt/naiscaper:${naiscaperVersion} /bin/bash -c \"/usr/bin/helm repo update && naiscaper ${clusterName} nais /root/nais-platform-apps\"")
-        }
-
-        stage("update nais 3rd party apps") {
-            sh """
-                if [[ -d ./nais-tpa/clusters/${clusterName} ]]; then
-                    docker run --rm -v `pwd`/nais-tpa:/root/nais-tpa -v `pwd`/${clusterName}:/root/.kube navikt/naiscaper:${naiscaperVersion} /bin/bash -c \"/usr/bin/helm repo update && /usr/bin/landscaper -v --env ${clusterName} --context ${clusterName} --namespace tpa apply --wait --wait-timeout 10m /root/nais-tpa/clusters/${clusterName}/*.yaml\"
-                else
-                    echo "No third party apps defined for ${clusterName}, skipping"
-                fi
-            """
-        }
-
-        stage("update nais platform apps (dry-run)") {
-            sh("""docker run --rm \
+            if (useBashscaper) {
+              sh """
+                  docker run --rm \
                     -v naiscaper-output:/naiscaper/output \
                     -v `pwd`/nais-platform-apps/base:/naiscaper/input/base:ro \
                     -v `pwd`/nais-platform-apps/clusters/${clusterName}:/naiscaper/input/overrides:ro \
                     navikt/naiscaper:${naiscaperDryRunVersion} \
-                    /bin/bash -c \"naiscaper /naiscaper/input/base /naiscaper/input/overrides /naiscaper/output\"""")
+                    /bin/bash -c \"naiscaper /naiscaper/input/base /naiscaper/input/overrides /naiscaper/output\"
+              """
 
-            sh("""docker run --rm \
+              sh """
+                  docker run --rm \
                     -v naiscaper-output:/apply \
                     -v `pwd`/${clusterName}:/root/.kube \
                     navikt/bashscaper:${bashscaperVersion} \
-                    /bin/bash -c \"/usr/bin/helm repo update && bashscaper nais ${clusterName} /apply/*.yaml\"""")
+                    /bin/bash -c \"/usr/bin/helm repo update && bashscaper nais ${clusterName} /apply/*.yaml\"
+              """
+            } else {
+              sh("docker run --rm -v `pwd`/nais-platform-apps:/root/nais-platform-apps -v `pwd`/${clusterName}:/root/.kube navikt/naiscaper:${naiscaperVersion} /bin/bash -c \"/usr/bin/helm repo update && naiscaper ${clusterName} nais /root/nais-platform-apps\"")
+            }
         }
 
-        stage("update nais 3rd party apps (dry-run)") {
-            sh """
-                if [[ -d ./nais-tpa/clusters/${clusterName} ]]; then
-                    docker run --rm \
-                      -v `pwd`/nais-tpa/clusters/${clusterName}:/apply \
-                      -v `pwd`/${clusterName}:/root/.kube \
-                      navikt/bashscaper:${bashscaperVersion} \
-                      /bin/bash -c \"/usr/bin/helm repo update && bashscaper tpa ${clusterName} /apply/*.yaml\"
-                else
-                    echo "No third party apps defined for ${clusterName}, skipping"
-                fi
-            """
+        stage("update nais 3rd party apps") {
+            if (useBashscaper) {
+              sh """
+                  if [[ -d ./nais-tpa/clusters/${clusterName} ]]; then
+                      docker run --rm \
+                        -v `pwd`/nais-tpa/clusters/${clusterName}:/apply \
+                        -v `pwd`/${clusterName}:/root/.kube \
+                        navikt/bashscaper:${bashscaperVersion} \
+                        /bin/bash -c \"/usr/bin/helm repo update && bashscaper tpa ${clusterName} /apply/*.yaml\"
+                  else
+                      echo "No third party apps defined for ${clusterName}, skipping"
+                  fi
+              """
+            } else {
+              sh """
+                  if [[ -d ./nais-tpa/clusters/${clusterName} ]]; then
+                      docker run --rm -v `pwd`/nais-tpa:/root/nais-tpa -v `pwd`/${clusterName}:/root/.kube navikt/naiscaper:${naiscaperVersion} /bin/bash -c \"/usr/bin/helm repo update && /usr/bin/landscaper -v --env ${clusterName} --context ${clusterName} --namespace tpa apply --wait --wait-timeout 10m /root/nais-tpa/clusters/${clusterName}/*.yaml\"
+                  else
+                      echo "No third party apps defined for ${clusterName}, skipping"
+                  fi
+              """
+            }
         }
 
         stage("deploy nais-testapp") {
